removed dep on tinygo sources;  basics of semihosting
Updated my code to use my proposed new "runtime" for
tinygo that allows runtimes to be defined externally.
I am using this mechanism, and will send them a PR
for including into the distribution for next release.

Added support for a single semihosting command (exit)
to prove that it works properly with qemu.